Extract SearchUsernameTypeahead into its own component
