+++ apps/app/src/client/components/Admin/AuditLog/AuditLogExportModal.tsx
@@ -4,13 +4,13 @@ import { useAtomValue } from 'jotai';
 import { useTranslation } from 'react-i18next';
 import { Modal, ModalBody, ModalFooter, ModalHeader } from 'reactstrap';
 
+import { SearchUsernameTypeahead } from '~/client/components/SearchUsernameTypeahead';
 import type { IAuditLogBulkExportRequestFilters } from '~/features/audit-log-bulk-export/interfaces/audit-log-bulk-export';
 import type { SupportedActionType } from '~/interfaces/activity';
 import { auditLogAvailableActionsAtom } from '~/states/server-configurations';
 
 import { DateRangePicker } from './DateRangePicker';
 import { DuplicateExportConfirmModal } from './DuplicateExportConfirmModal';
-import { SearchUsernameTypeahead } from './SearchUsernameTypeahead';
 import { SelectActionDropdown } from './SelectActionDropdown';
 import { useAuditLogExport } from './useAuditLogExport';
