+++ apps/app/src/client/components/SearchUsernameTypeahead/index.ts
@@ -1,0 +1,1 @@
+export { SearchUsernameTypeahead } from './SearchUsernameTypeahead';
